fix: :bug: Handle 0 decimal place rounding for bits_to_name
There appears to be a quirk with the round method that it will show the number as a float if you specify 0 decimal places and as an int if you don't specify the places at all. This tweak should fix the issue.

refactor: :recycle: Simplify code to reduce code duplication