docs: Added ignore for pydocstyle
Using escaped quotes in pydocs is, apparently, frowned upon and flagged by pydocstyle D300 and D301. However, in the ip_classfull example, the escaped double quotes are for triple double quotes. Removing the backslashs and making the pydoc a raw string won't work, the backslashes still have to be there. Therefore, we're ignoring D300 and D301 in this case.